mutation_fn_t is now value-returning
in my recent update to the mutation function in run_phonopy,
I repeated my earlier mistake of forgetting to assign the new
value back in-place to the argument.

surprisingly, no compiler warnings were generated for the erroneous
'return x' in what should have been a void function, making this bug
fairly difficult to track down each time I've made it.

The only solution is to prevent it from the get-go.