fractional coordinate utils on structure_t
apparently colin has never needed to work much with fractional
coords.  However, for a recent addition I desparately needed
the ability to preserve fractional coordinates while scaling
the lattice, so I added a core of utility functions to enable this.

this also fixes what was apparently unused code for a 3x3
matrix inverse