fix warnings and memory leak
So it turns out Py_BuildValue does perform an INCREF on 'O' values.
When I "fixed" this I was probably mistakenly reading the docs for
the PyArg family of functions, which are just a bit higher on the
same page...

one can also clearly see that `to_python(std::tuple)` has not
yet been used.